Finished adding events

+++ dotNet5782_8871_6445/PL/StationPage.xaml.cs
@@ -103,10 +103,16 @@ namespace PL
 
         private void UpdateDeleteStationButton_Click(object sender, RoutedEventArgs e)
         {
+            MessageBoxResult res = MessageBox.Show("Are you sure you want to delete selected station?", "Verification", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+            if (res == MessageBoxResult.No)
+                return;
+            MessageBox.Show("This method is under construction!", "TBD", MessageBoxButton.OKCancel, MessageBoxImage.Asterisk);
+            MessageBox.Show("This method is under construction!", "TBD", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+            MessageBox.Show("This method is under construction!", "TBD", MessageBoxButton.YesNo, MessageBoxImage.Stop);
+            MessageBox.Show("This method is under construction!", "TBD", MessageBoxButton.YesNoCancel, MessageBoxImage.Hand);
             try
             {
                 IBL.DeleteStation(StationBL.ID);
-                NavigationService.GoBack();
             }
             catch (Exception exp)
             {
